backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
asyncpg==0.30.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
nltk==3.9.1
pandas==2.2.3
scikit-learn==1.6.1
firebase-admin==6.6.0
pytest==8.3.4
pytest-asyncio==0.25.0
